// app/shared/enums/urls.enum.ts
export enum Urls {
  ACCESS_TOKEN_UPDATE = 'api/account/updateaccesstoken',
  APP_SETTINGS = 'api/settings/appsettings',
  AREA_COVERED = 'api/post/areacovered',
  AUTH_CODE = 'api/account/requestaccesstoken',
  EULA = 'api/put/eulaaccepted',
  FAULT = 'api/get/faults',
  LOGIN = 'api/account/login',
  USER_DETAILS = 'api/get/userprofile',
  USER_PREFERENCE = 'api/userpreference',
  VEHICLE_ICON = './assets/Icon-Vehicle',
  USER_DATA = '../account/getuserprofile',
  DATA_AUTH_CODE = 'api/account/requestdataaccesstoken',
  LAST_CREATED = 'api/get/GetLastCreated',
  GET_FILEDATA = 'api/get/GetFiledata',
  GET_CLIENTS = 'api/get/GetClients',
  GET_CLIENTS_CONNECT_INFO = 'api/get/GetClientsConnectInfo',
  GET_INVITE_ID = 'api/get/getInviteId',
  GET_FIELDS = 'api/get/GetFields',
  SHARES_ME = 'api/get/SharesMe',
  GET_USER_DATA = 'api/get/userprofile',
  GET_MKT_CLIENTS_CONNECT_INFO = 'api/get/getMktClientsConnectInfo',
  GET_FIELD_DATA = 'api/get/getfieldData',
  DELETE_SHARE = 'api/delete/deleteShare',
  DISCONNECT_APP = 'api/delete/DisconnectApp',
  PATCH_SHARE = 'api/patch/PatchShare',
  CREATE_SHARE = 'api/post/CreateShare',
  RESEND_INVITE = 'api/put/resendinvite',
  VEHICLE_CHART_DATA = 'api/get/duty',
  VEHICLE_FUEL_DATA = 'api/get/fuel',
  CUSTOM_P_ICON = './assets/customparameters-images/',
  SHARE_INVITE = 'api/post/DataAccess',
  CLIMATE_ACCESS_TOKEN = 'api/get/GetClimateAccessToken',
  CLIMATE_SAVETOKEN = 'api/post/SaveClimateToken',
  DISCONNECTCLIMATE_APP = 'api/delete/DisconnectClimate',
  GETSTATUSOFUSER = 'api/get/GetStatusOfUser',
  CLIMATEEXTURL = 'https://climate.com/static/app-login/index.html?scope=exports:read+fields:read+diagnostics:read&page=oidcauthn&response_type=code&redirect_uri=',
  ENABLE_CLIMATE = 'api/post/EnableClimate',
  DISABLE_CLIMATE = 'api/post/DisableClimate',
  FILE_NOTIFICATIONS = 'api/get/Notifications',
  UPLOAD_TASK_FILE = 'api/post/UploadTaskFile',
  USER_NOTIFICATIONS = 'api/fleet/get/GetNotifications',
  SEND_INBOX_FILE_TRANSFER = 'api/post/SendFile',
  GET_FMIS_FILTERS = 'api/get/GetFMISFilters',
  GET_EULA = 'api/post/eula'
}